lodash add
