+++ packages/ui/src/RSwitch/index.tsx
@@ -3,7 +3,7 @@ import { Switch } from 'tamagui';
 const RSwitch = (props) => {
   return (
     <Switch {...props}>
-      <Switch.Thumb/>
+      <Switch.Thumb />
     </Switch>
   );
 };
